Adapt sidebar marker icon to report status
Implemented by factoring out the getImageSrc util used for the marker pins
