+++ src/pages/Reports/pages/OverviewMap/components/ClusteredMarkers/marker-utils.js
@@ -1,33 +1,6 @@
 import MapboxGL from 'mapbox-gl';
 
-// import MarkerNew from '~/images/reports/pin-new.png';
-import MarkerVerification from '~/images/reports/pina-meldung@2x.png';
-import MarkerAccepted from '~/images/reports/pin-accepted.png';
-import MarkerRejected from '~/images/reports/pin-rejected.png';
-import MarkerDone from '~/images/reports/pin-done.png';
-
-import logger from '~/utils/logger';
-
-// The "new" marker is not used anymore and replaced by the verification marker
-// for now
-const ICONS_BY_STATUS = {
-  new: MarkerVerification,
-  verification: MarkerVerification,
-  accepted: MarkerAccepted,
-  rejected: MarkerRejected,
-  done: MarkerDone
-};
-
-/**
- * Returns a marker image source depending on report status
- *
- * @param {Object} markerData containing the status field
- */
-const getMarkerSrc = (markerData) => {
-  const { status } = markerData;
-  if (status == null) logger('Record is missing status:', markerData);
-  return ICONS_BY_STATUS[status];
-};
+import utils from '~/pages/Reports/utils';
 
 function createClusterMarker({ pointCount, map, clusterSource, id, lngLat }) {
   const el = document.createElement('div');
@@ -90,7 +63,7 @@ function createPinMarker({
 
   const updatedMarkerData = { ...markerData, geometry, details };
 
-  el.innerHTML = `<img class="marker-image" src="${getMarkerSrc(
+  el.innerHTML = `<img class="marker-image" src="${utils.getMarkerSrc(
     markerData
   )}" />`;
   el.addEventListener('click', (evt) => onClick(evt, updatedMarkerData));
